Remove inputtest class and rename to media util
