--- src/main/java/FileInput/FileInputTest.java
@@ -1,32 +1,0 @@
-package FileInput;
-
-//import ca.ciccc.silverBullet.gameplayEnums.PlayerAction;
-//import ca.ciccc.silverBullet.gridElements.GridBoard;
-import javafx.animation.AnimationTimer;
-import javafx.application.Application;
-import javafx.scene.Node;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.layout.Pane;
-import javafx.stage.Stage;
-
-import javafx.scene.media.Media;
-import javafx.scene.media.MediaPlayer;
-import java.io.File;
-import java.net.URL;
-
-import java.util.Collections;
-import java.util.List;
-
-/**
- * A very simple Java FX Application used to demonstrate a couple of aspects of the Java 11
- * ecosystem.
- */
-public class FileInputTest extends Application {
-  @Override
-  public void start(Stage stage) throws Exception {
-
-  }
-}
-
-
