+++ src/main/java/ca/ciccc/silverBullet/SilverBulletApp.java
@@ -1,11 +1,10 @@
 package ca.ciccc.silverBullet;
 
-import FileInput.FileInput;
+import ca.ciccc.silverBullet.utils.MediaUtil;
 import ca.ciccc.silverBullet.controller.MenuController;
 import ca.ciccc.silverBullet.utils.ConstUtil;
 import java.util.ResourceBundle;
 import javafx.application.Application;
-import javafx.scene.media.AudioClip;
 import javafx.stage.Stage;
 
 /**
@@ -14,7 +13,7 @@ import javafx.stage.Stage;
 public class SilverBulletApp extends Application {
 
   public static Stage primaryStage;
-  FileInput fileInput = new FileInput();
+  MediaUtil mediaUtil = new MediaUtil();
   public static void main(String[] args) {
     ConstUtil.setResourceBundle(ResourceBundle.getBundle("application"));
     launch();
