+++ src/main/java/ca/ciccc/silverBullet/playerElements/Player.java
@@ -1,6 +1,6 @@
 package ca.ciccc.silverBullet.playerElements;
 
-import FileInput.FileInput;
+import ca.ciccc.silverBullet.utils.MediaUtil;
 import ca.ciccc.silverBullet.enums.gameplay.Directions;
 import ca.ciccc.silverBullet.enums.gameplay.Orientation;
 import ca.ciccc.silverBullet.enums.gameplay.PlayerAction;
@@ -15,7 +15,7 @@ import javafx.scene.shape.Circle;
 public class Player {
 
     private boolean hasShot;
-    private FileInput fileInput = new FileInput();
+    private MediaUtil mediaUtil = new MediaUtil();
     private int playerNumber;
     private Directions facingDirection;
     private int gridPositionX;
@@ -51,10 +51,10 @@ public class Player {
         Image image = null;
         if(playerNumber == 1)
         {
-            image = fileInput.image("File:src/main/resources/images/Character/Fire/Fire.png");
+            image = mediaUtil.image("File:src/main/resources/images/Character/Fire/Fire.png");
         }else if(playerNumber == 2)
         {
-            image = fileInput.image("File:src/main/resources/images/Character/Rock/Rock.png");
+            image = mediaUtil.image("File:src/main/resources/images/Character/Rock/Rock.png");
         }
         ((Circle) playerNode).setFill(new ImagePattern(image));
         playerActionCounter = new ActionCounter(playerNumber);
@@ -281,7 +281,7 @@ public class Player {
         }
 
         String path = "File:src/main/resources/images/Character/" + playerelement +"/" + playerelement + dirction +".png";
-        Image img = fileInput.image(path);
+        Image img = mediaUtil.image(path);
         ((Circle) this.playerNode).setFill(new ImagePattern(img));
     }
 }
